display: crt=on is a soft 4:3 tube, plus a vintage terminal look
Rework SET DISPLAY crt=on from hard scan lines to the 4:3 tube stretch
softened by LINEAR filtering, so the vertical stretch reads as a smooth
phosphor face rather than blocky row duplication. The crisp default keeps
NEAREST so low-res pixels stay square. The old soft-scanline overlay
(Window::scanTex, ensureScanTex) is removed.

The built-in terminal now ships the authentic DEC VT220 face
(boards/terminal-vt220font.h, a 10x20 cell) in place of the borrowed
VDM-1 CGEN, so glyphs have real inter-character spacing. TerminalFont::
glyphRow() widens to uint16_t with bit 15 the leftmost dot, so a cell may
be wider than eight dots; eight-dot fonts MSB-align into the top of the
word.

New terminal connect-string options: phosphor=green|amber picks the tube
colour, width=PX opens the window at an exact pixel width. The exact-width
fill is gated to the crt look, so the crisp default keeps integer scaling.
Toned the default green palette down so it is easier on the eyes.

examples/cpm/cpm22-terminal.toml turns the crt look on and opens the
terminal at width=1100. Manual (configuring, serial, boards), devguide
(terminal), monitor help and the changelog updated for the new knobs.

+++ src/host/terminal/font.h
@@ -7,8 +7,10 @@
 // descender flag), a VT100 has its own ROM, a plain build ships a bundled font. So
 // the renderer draws against this interface and each terminal supplies the metal.
 //
-// glyphRow() returns one scan line of a glyph with bit 7 = the LEFTMOST dot, which is
-// the convention the VDM-1 and VDB share; a set bit paints the foreground.
+// glyphRow() returns one scan line of a glyph with bit 15 = the LEFTMOST dot; a set bit
+// paints the foreground. The value is 16 bits so a cell may be wider than 8 dots -- the
+// authentic VT220 face is 10 dots wide. An 8-dot font (the VDM-1, the VDB) MSB-aligns its
+// byte into the top of the word, so bit 15 stays the leftmost dot for every font.
 
 #include <cstdint>
 
@@ -21,8 +23,8 @@ public:
     virtual int cellCols() const = 0;  // dot columns per cell (the glyph's box width)
     virtual int cellRows() const = 0;  // scan lines per cell (the glyph's box height)
 
-    // One scan line (0..cellRows()-1) of glyph `code`, bit 7 = leftmost dot.
-    virtual uint8_t glyphRow(uint8_t code, int row) const = 0;
+    // One scan line (0..cellRows()-1) of glyph `code`, bit 15 = leftmost dot.
+    virtual uint16_t glyphRow(uint8_t code, int row) const = 0;
 };
 
 } // namespace altair
